Split up BatchStoreTest
Split `BatchStoreTest` into smaller pieces following the pattern in commit 6025548.
