+++ src/test/kotlin/com/terraformation/backend/nursery/db/batchStore/BatchStoreUpdateDetailsTest.kt
@@ -1,0 +1,66 @@
+package com.terraformation.backend.nursery.db.batchStore
+
+import com.terraformation.backend.db.nursery.BatchId
+import com.terraformation.backend.db.nursery.tables.pojos.BatchesRow
+import com.terraformation.backend.nursery.db.BatchStaleException
+import io.mockk.every
+import java.time.Instant
+import java.time.LocalDate
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
+
+internal class BatchStoreUpdateDetailsTest : BatchStoreTest() {
+  private val batchId = BatchId(1)
+  private val updateTime = Instant.ofEpochSecond(1000)
+
+  @BeforeEach
+  fun setUpTestBatch() {
+    insertBatch(
+        BatchesRow(notes = "initial notes", readyByDate = LocalDate.EPOCH),
+        id = batchId,
+        readyQuantity = 1,
+        speciesId = speciesId)
+
+    every { clock.instant() } returns updateTime
+  }
+
+  @Test
+  fun `updates values`() {
+    val before = batchesDao.fetchOneById(batchId)!!
+
+    store.updateDetails(
+        batchId = batchId, version = 1, notes = "new notes", readyByDate = LocalDate.of(2022, 1, 1))
+
+    val after = batchesDao.fetchOneById(batchId)!!
+
+    assertEquals(
+        before.copy(
+            notes = "new notes",
+            modifiedTime = updateTime,
+            readyByDate = LocalDate.of(2022, 1, 1),
+            version = 2),
+        after)
+  }
+
+  @Test
+  fun `can set optional values to null`() {
+    val before = batchesDao.fetchOneById(batchId)!!
+
+    store.updateDetails(batchId = batchId, version = 1, notes = null, readyByDate = null)
+
+    val after = batchesDao.fetchOneById(batchId)!!
+
+    assertEquals(
+        before.copy(notes = null, modifiedTime = updateTime, readyByDate = null, version = 2),
+        after)
+  }
+
+  @Test
+  fun `throws exception if version number does not match current version`() {
+    assertThrows<BatchStaleException> {
+      store.updateDetails(batchId = batchId, version = 0, notes = null, readyByDate = null)
+    }
+  }
+}
